Names changed a bit.

+++ src/fiflo.h
@@ -7,11 +7,11 @@
 #include <stdint.h>
 
 #define BUFF_SZ SHRT_MAX - 1
-typedef int16_t buff_t; // > 16 makes stack error.
+typedef int32_t buff_t; // > 16 makes stack error. TODO
 
 struct Data // There is no typedef to provide better code readibility.
 {
-	char *filename; // Eg. /home/user/basename.asdf
+	char *filename; // Eg. /home/user/basename
 	char *text;
 	buff_t chars;
 	buff_t lines;
